+++ app/src/navigation/router.tsx
@@ -59,6 +59,7 @@ import SupportScreen from "@/scenes/support/SupportScreen";
 import DiaryDetail from "@/scenes/status/DiaryDetail";
 import { ModalTriggerScreen } from "@/scenes/suivi/triggers/ModalTrigger";
 import SettingsModal from "@/scenes/settings/settings-modal";
+import CommityScreen from "@/scenes/commity/CommityScreen";
 
 const Stack = createStackNavigator();
 
@@ -264,6 +265,7 @@ class Router extends React.Component<RouterProps> {
               }}
             />
             <Stack.Screen name="settings" component={SettingsModal} />
+            <Stack.Screen name="commity" component={CommityScreen} />
             <Stack.Screen name="indicators-settings-more" component={IndicatorsSettingsMore} />
             <Stack.Screen name="goals-settings" component={GoalsSettings} />
             <Stack.Screen name="goals-settings-more" component={GoalsSettingsMore} />